+++ dipy/tracking/local/localtracking.py
@@ -1,8 +1,14 @@
 import numpy as np
 
 from .localtrack import local_tracker
+from dipy.align import Bunch
 from dipy.tracking import utils
-from dipy.tracking.local import TissueTypes
+
+# enum TissueClass (tissue_classifier.pxd) is not accessible
+# from here. To be changed when minimal cython version > 0.21.
+# cython 0.21 - cpdef enum to export values into Python-level namespace
+# https://github.com/cython/cython/commit/50133b5a91eea348eddaaad22a606a7fa1c7c457
+TissueTypes = Bunch(OUTSIDEIMAGE=-1, INVALIDPOINT=0, TRACKPOINT=1, ENDPOINT=2)
 
 
 class LocalTracking(object):
